add Charts class
