+++ src/app/components/template/chord/chord.component.ts
@@ -1,24 +1,19 @@
 import { AfterViewInit, Component, Input, OnChanges, SimpleChanges } from '@angular/core';
-import * as echarts from 'echarts';
-import { isNil } from 'lodash';
 import { EChartsOption } from 'echarts';
-import { ChordCategory, ChordLink, ChordNode } from 'src/app/types/charts';
+import { Charts, ChordCategory, ChordLink, ChordNode } from 'src/app/types/charts';
 
 @Component({
   selector: 'app-chord',
   templateUrl: './chord.component.html',
   styleUrls: ['./chord.component.scss']
 })
-export class ChordComponent implements AfterViewInit, OnChanges {
+export class ChordComponent extends Charts implements AfterViewInit, OnChanges {
 
   @Input() categories: ChordCategory[] = [];
   @Input() links: ChordLink[] = [];
   @Input() nodes: ChordNode[] = [];
   @Input() useForce: boolean = false;
   @Input() id: string;
-
-  private chart: echarts.ECharts | undefined;
-  private option: EChartsOption;
 
   ngAfterViewInit() {
     this.updateChart();
@@ -34,21 +29,11 @@ export class ChordComponent implements AfterViewInit, OnChanges {
   }
 
   constructor() {
+    super();
     this.id = "";
-    this.chart = undefined;
-    this.option = {};
   }
 
-  private updateChart(): void {
-    const element = document.getElementById(this.id);
-    if (isNil(this.chart) && !isNil(element)) {
-      this.chart = echarts.init(document.getElementById(this.id)!);
-    }
-    this.option = this.generateChartOption();
-    this.option && this.chart?.setOption(this.option);
-  }
-
-  private generateChartOption(): EChartsOption {
+  generateChartOption(): EChartsOption {
     return {
       tooltip: {},
       legend: [
@@ -70,7 +55,6 @@ export class ChordComponent implements AfterViewInit, OnChanges {
           data: this.nodes,
           edges: this.links,
           categories: this.categories,
-          // roam: true,
           draggable: true,
           label: {
             show: true,
@@ -87,12 +71,6 @@ export class ChordComponent implements AfterViewInit, OnChanges {
           },
           lineStyle: {
             color: 'source'
-          },
-          force: {
-            // repulsion: 50,
-            // edgeLength: 5,
-            // repulsion: 20,
-            // gravity: 0.2
           }
         }
       ]
