+++ src/app/components/template/heatmap/heatmap.component.ts
@@ -1,14 +1,12 @@
 import { AfterViewInit, Component, Input, OnChanges, SimpleChanges } from '@angular/core';
-import * as echarts from 'echarts';
-import { isNil } from 'lodash';
-import { EChartsOption, HeatmapData } from 'src/app/types/charts';
+import { Charts, EChartsOption, HeatmapData } from 'src/app/types/charts';
 
 @Component({
   selector: 'app-heatmap',
   templateUrl: './heatmap.component.html',
   styleUrls: ['./heatmap.component.scss']
 })
-export class HeatmapComponent implements AfterViewInit, OnChanges  {
+export class HeatmapComponent extends Charts implements AfterViewInit, OnChanges  {
 
   @Input() data: HeatmapData[] = [];
   @Input() xAxis: string[] = [];
@@ -16,9 +14,6 @@ export class HeatmapComponent implements AfterViewInit, OnChanges  {
   @Input() id: string;
 
   private max = -1;
-
-  private chart: echarts.ECharts | undefined;
-  private option: EChartsOption;
 
   ngAfterViewInit() {
     this.updateChart();
@@ -37,21 +32,11 @@ export class HeatmapComponent implements AfterViewInit, OnChanges  {
   }
 
   constructor() {
+    super();
     this.id = "";
-    this.chart = undefined;
-    this.option = {};
   }
 
-  private updateChart(): void {
-    const element = document.getElementById(this.id);
-    if (isNil(this.chart) && !isNil(element)) {
-      this.chart = echarts.init(document.getElementById(this.id)!);
-    }
-    this.option = this.generateChartOption();
-    this.option && this.chart?.setOption(this.option);
-  }
-
-  private generateChartOption(): EChartsOption {
+  generateChartOption(): EChartsOption {
     return {
       tooltip: {},
       grid: {
