+++ src/app/components/template/histogram/histogram.component.ts
@@ -1,7 +1,5 @@
 import { AfterViewInit, Component, Input, OnChanges, SimpleChanges } from '@angular/core';
-import * as echarts from 'echarts';
-import { isNil } from 'lodash';
-import { EChartsOption } from 'src/app/types/charts';
+import { Charts, EChartsOption } from 'src/app/types/charts';
 
 const GRADE_VALUES = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]; // TODO : give the values to count with Input
 
@@ -10,13 +8,10 @@ const GRADE_VALUES = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]; // TODO : give the values 
   templateUrl: './histogram.component.html',
   styleUrls: ['./histogram.component.scss']
 })
-export class HistogramComponent implements AfterViewInit, OnChanges {
+export class HistogramComponent extends Charts implements AfterViewInit, OnChanges {
 
   @Input() values: number[] = [];
   @Input() id: string;
-
-  private chart: echarts.ECharts | undefined;
-  private option: EChartsOption;
 
   ngAfterViewInit() {
     this.updateChart();
@@ -27,22 +22,12 @@ export class HistogramComponent implements AfterViewInit, OnChanges {
     this.updateChart();
   }
 
-  constructor() {
+  constructor() { 
+    super();
     this.id = "";
-    this.chart = undefined;
-    this.option = {};
   }
 
-  private updateChart(): void {
-    const element = document.getElementById(this.id);
-    if (isNil(this.chart) && !isNil(element)) {
-      this.chart = echarts.init(document.getElementById(this.id)!);
-    }
-    this.option = this.generateChartOption();
-    this.option && this.chart?.setOption(this.option);
-  }
-
-  private generateChartOption(): EChartsOption {
+  generateChartOption(): EChartsOption {
     return {
       tooltip: {
         trigger: 'item'
